Refactor repositories and entities for flexibility
Refactored repository methods to enhance querying, deletion, and
update capabilities. Introduced `FindBy` for conditional queries
with eager loading. Removed `GetAllAsync` in favor of more
specific methods like `GetAllBy` and `GetAll`. Added `DeleteAsync`
to multiple repositories with null-check handling.

Updated entities to align with new repository patterns. Removed
`IsDeleted` and `ProcessedAt` from `Payment`. Added `UpdatedAt`
to entities implementing `IUpdatable`. Simplified soft deletion
and timestamp handling using `IUpdatable` and `IDeletable`
interfaces in `BaseEntityInterceptor`.

Refactored configurations to remove unnecessary constraints and
align default values with `SqlCommon.Date`. Improved consistency
in method signatures, namespaces, and repository patterns.

+++ Dentizone.Application/Repositories/PostRepsitory.cs
@@ -1,4 +1,5 @@
-﻿using Dentizone.Application.Abstracts;
+﻿using System.Linq.Expressions;
+using Dentizone.Application.Abstracts;
 using Dentizone.Application.Interfaces;
 using Dentizone.Domain.Entity;
 using Dentizone.Infrastructure;
@@ -19,9 +20,29 @@ namespace Dentizone.Application.Repositories
             return entity;
         }
 
+        public async Task<Post?> FindBy(Expression<Func<Post, bool>> condition,
+            Expression<Func<Post, object>>[]? includes)
+        {
+            IQueryable<Post> query = dbContext.Posts;
+            if (includes != null)
+            {
+                foreach (var include in includes)
+                {
+                    query = query.Include(include);
+                }
+            }
+
+            return await query.FirstOrDefaultAsync(condition);
+        }
+
         public async Task<Post?> DeleteAsync(string id)
         {
             var post = await GetByIdAsync(id);
+
+            if (post == null)
+            {
+                return null;
+            }
 
 
             dbContext.Posts.Remove(post);
@@ -36,6 +57,32 @@ namespace Dentizone.Application.Repositories
                 .Skip(skippedPages)
                 .Take(DefaultPageSize)
                 .ToListAsync();
+        }
+
+        public async Task<IEnumerable<Post>> GetAllAsync(int page, Expression<Func<Post, bool>>? filter,
+            Expression<Func<Post, object>>? orderBy, Expression<Func<Post, object>>[]? includes = null)
+        {
+            IQueryable<Post> query = dbContext.Posts;
+            if (filter != null)
+            {
+                query = query.Where(filter);
+            }
+
+            if (includes != null)
+            {
+                foreach (var include in includes)
+                {
+                    query = query.Include(include);
+                }
+            }
+
+            if (orderBy != null)
+            {
+                query = query.OrderBy(orderBy);
+            }
+
+            var skippedPages = CalculatePagination(page);
+            return await query.Skip(skippedPages).Take(DefaultPageSize).ToListAsync();
         }
 
         public async Task<Post?> GetByIdAsync(string id)
